% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINSEG.R
\name{BINSEG}
\alias{BINSEG}
\title{Binary Segmentation}
\usage{
BINSEG(data.vec, Kmax)
}
\arguments{
\item{data.vec}{input data vector, [number of data points]}

\item{Kmax}{max number of segments}
}
\value{
Vector of square loss values, [number of segments]
}
\description{
Binary segmentation for changepoint detection on a sequential data set up to K segments
}
\examples{
data(neuroblastoma, package="neuroblastoma")
library(data.table)
nb.dt <- data.table(neuroblastoma$profiles)
data.dt <- nb.dt[profile.id=="1" & chromosome=="1"]
BINSEG(data.dt$logratio,3)
}
