% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DYNPROG_interface}
\alias{DYNPROG_interface}
\title{Dynamic programming for optimal segmentation}
\usage{
DYNPROG_interface(data_vec, N_segments)
}
\arguments{
\item{data_vec}{numeric data vector, [N_data]}

\item{N_segments}{number of max segments}
}
\value{
loss_mat, loss matrix [N_data] x [N_segments]
}
\description{
Dynamic programming for changepoint detection of a givens sequence of data with up to N_segments
}
\examples{
data(neuroblastoma, package="neuroblastoma")
library(data.table)
nb.dt <- data.table(neuroblastoma$profiles)
data.dt <- nb.dt[profile.id=="1" & chromosome=="1"]
DYNPROG_interface(data.dt$logratio,3)
}
